Move find_top_3_insertions function and structs to a dedicated header.
